rename to JSX and move views to view folder
