+++ src/views/Stake.jsx
