+++ src/views/index.js
@@ -1,3 +1,10 @@
 export { default as ExampleUI } from "./ExampleUI";
 export { default as Hints } from "./Hints";
 export { default as Subgraph } from "./Subgraph";
+
+// scaffold examples above. 
+export { default as Bond } from './Bond'
+export { default as ChooseBond } from './ChooseBond';
+export { default as Stake } from './Stake';
+export { default as Dashboard } from './Dashboard';
+
